requirements.txt
django
djangorestframework
whitenoise
Pillow
django-imagekit